YASP: Updated consolidation algorithm
The previous algorithm of calling the accoustic model directly
yielded unreliable results. The new algorithm uses the standard
ps_start_utt(), then grab the alignment data. The alignment data
is relatively correct to the timing information in the word hypothesis
but it needs to be offset to start with the words.

Signed-off-by: Amir Shehata <amir.shehata@gmail.com>